# services/event-service/docs/swagger.yaml
basePath: /api/v1
definitions:
  event-service_internal_dto.CreateEventRequest:
    properties:
      category:
        type: string
      title:
        type: string
    type: object
  event-service_internal_dto.CreateEventResponse:
    properties:
      category:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  event-service_internal_dto.EventListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/event-service_internal_dto.EventResponse'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  event-service_internal_dto.EventResponse:
    properties:
      category:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  event-service_internal_dto.UpdateEventRequest:
    properties:
      category:
        type: string
      title:
        type: string
    type: object
  internal_handler.APIError:
    properties:
      Code:
        type: string
      details: {}
      message:
        type: string
      request_id:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: backend@iticket.io
    name: Backend Team
  description: Event service for Iticket platform
  termsOfService: http://example.com/terms/
  title: Event Service API
  version: "1.0"
paths:
  /events:
    get:
      description: Get paginated list of events
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Category filter
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event-service_internal_dto.EventListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.APIError'
      summary: List events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Create a new event
      parameters:
      - description: Event payload
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/event-service_internal_dto.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/event-service_internal_dto.CreateEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.APIError'
      summary: Create event
      tags:
      - events
  /events/{id}:
    delete:
      description: Delete event by ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.APIError'
      summary: Delete event
      tags:
      - events
    put:
      consumes:
      - application/json
      description: Update event by ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/event-service_internal_dto.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event-service_internal_dto.EventResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.APIError'
      summary: Update event
      tags:
      - events
swagger: "2.0"